packages/inference/requirements.txt
# Davo-AI · Inference servers requirements
# Python 3.11+

# ─── Core ML ────────────────────────────────────────────
vllm>=0.7.0                       # LLM + Vision serving (OpenAI-compatible API)
faster-whisper>=1.0.0             # Whisper STT с CTranslate2
ctranslate2>=4.4.0
ultralytics>=8.3.0                # YOLOv8

# ─── PyTorch (CUDA 12.4) ────────────────────────────────
torch>=2.5.0
torchvision>=0.20.0

# ─── HuggingFace ─────────────────────────────────────────
transformers>=4.45.0
accelerate>=1.0.0
huggingface-hub>=0.25.0
sentencepiece>=0.2.0

# ─── Web servers ─────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic>=2.9.0
httpx>=0.27.0

# ─── Image / Video ───────────────────────────────────────
pillow>=10.4.0
opencv-python-headless>=4.10.0
numpy>=1.26.0,<2.0.0

# ─── Logging / Utils ─────────────────────────────────────
loguru>=0.7.2
rich>=13.9.0
